requirements.txt
numpy
scipy
sounddevice
soundfile
PyQt6
pyyaml
mutagen
pyinstaller
